Lowering the chain limit to 545
Summing the first few primes until the sum exceeds the limit of 1M gives
the upper bound on the length of the chain of 545. The sum of the first
546 primes is 1001602. The sum of the first 545 primes is 997659.  Of
course the length of the chain has to be odd, otherwise the sum would be
even, and hence not a prime.